Retrieve pathway topology as weighted adjacent matrix

Retrieve pathway topology matrices and convert to normalized weighted directed adjacency matrices describing gene signaling networks.

retrieve_topology(database, pathwayName = NULL, beta = NULL)

Arguments

database

See example for supported databases.

pathwayName

Optional. Subset of pathway names as a vector.

beta

Optional. A named numeric vector of weights to be assigned to each type of gene/protein relation type. See details for more information.

Value

A list where each element is a matrix corresponding to a pathway

Details

This function takes the pathway topology information retrieved using graphite and convert them to normalized weighted directed adjacency matrices describing the gene signaling network, which can be directed used to compute pathway perturbation score through SPIA algorithm. See cited document for more details.

The beta parameter specifies weights to be assigned to each type of gene-gene interaction. It should be a named numeric vector of length 23, whose names must be: c("activation","compound","binding/association","expression","inhibition","activation_phosphorylation","phosphorylation", "indirect","inhibition_phosphorylation","dephosphorylation_inhibition","dissociation","dephosphorylation","activation_dephosphorylation", "state","activation_indirect","inhibition_ubiquination","ubiquination","expression_indirect","indirect_inhibition", "repression","binding/association_phosphorylation","dissociation_phosphorylation","indirect_phosphorylation"). If unspecified, beta will be by default chosen as: c(1,0,0,1,-1,1,0,0,-1,-1,0,0,1,0,1,-1,0,1,-1,-1,0,0,0).

The converted weighted adjacent matrices will be stored in a list. We recommend users to store the returned list as a file so this step only needs to be performed once.

This function only supports and can only be used to retreive human databases as this stage.

References

Tarca AL, Draghici S, Khatri P, Hassan SS, Mittal P, Kim JS, Kim CJ, Kusanovic JP, Romero R. A novel signaling pathway impact analysis. Bioinformatics. 2009 Jan 1;25(1):75-82. Sales, G., Calura, E., Cavalieri, D. et al. graphite - a Bioconductor package to convert pathway topology to gene network. BMC Bioinformatics 13, 20 (2012).

Examples

# explore all databases supported by graphite
graphite::pathwayDatabases()
#>          species     database
#> 1      athaliana         kegg
#> 2      athaliana     pathbank
#> 3      athaliana wikipathways
#> 4        btaurus         kegg
#> 5        btaurus     pathbank
#> 6        btaurus     reactome
#> 7        btaurus wikipathways
#> 8       celegans         kegg
#> 9       celegans     pathbank
#> 10      celegans     reactome
#> 11      celegans wikipathways
#> 12   cfamiliaris         kegg
#> 13   cfamiliaris     reactome
#> 14   cfamiliaris wikipathways
#> 15 dmelanogaster         kegg
#> 16 dmelanogaster     pathbank
#> 17 dmelanogaster     reactome
#> 18 dmelanogaster wikipathways
#> 19        drerio         kegg
#> 20        drerio     reactome
#> 21        drerio wikipathways
#> 22         ecoli         kegg
#> 23         ecoli     pathbank
#> 24         ecoli wikipathways
#> 25       ggallus         kegg
#> 26       ggallus     reactome
#> 27       ggallus wikipathways
#> 28      hsapiens         kegg
#> 29      hsapiens      panther
#> 30      hsapiens     pathbank
#> 31      hsapiens     pharmgkb
#> 32      hsapiens     reactome
#> 33      hsapiens        smpdb
#> 34      hsapiens wikipathways
#> 35     mmusculus         kegg
#> 36     mmusculus     pathbank
#> 37     mmusculus     reactome
#> 38     mmusculus wikipathways
#> 39   rnorvegicus         kegg
#> 40   rnorvegicus     pathbank
#> 41   rnorvegicus     reactome
#> 42   rnorvegicus wikipathways
#> 43   scerevisiae         kegg
#> 44   scerevisiae     pathbank
#> 45   scerevisiae     reactome
#> 46   scerevisiae wikipathways
#> 47       sscrofa         kegg
#> 48       sscrofa     reactome
#> 49       sscrofa wikipathways
#> 50       xlaevis         kegg
gsTopology <- retrieve_topology(database = "kegg")
# if only interested in selected pathways, specify the pathway names in the `pathwayName` parameter
gsTopology <- retrieve_topology(database = "kegg",
pathwayName = c("Glycolysis / Gluconeogenesis",
"Citrate cycle (TCA cycle)","Pentose phosphate pathway"))