Compute weighted single sample LogFCs from normalised logCPM

Compute weighted single sample logFCs for each treated samples using normalised logCPM values. Fit a lowess curve on variance of single sample logFCs ~ mean of logCPM, and use it to predict a gene-wise weight. The weighted single sample logFCs are ready for computing perturbation scores.

weight_ssFC(logCPM, metadata, factor, control)

Arguments

logCPM

Matrix of normaslised logCPM where rows are genes and columns are samples. Row names need to be gene entrez ID and column names need to be sample names

metadata

Sample metadata data frame as described in the details section.

factor

Factor defines how samples can be put into matching pairs (eg. patient).

control

Treatment level that is the control.

Value

A list with two elements: $weight gene-wise weights; $logFC weighted single sample logFC matrix

Details

This function computes weighted single sample logFCs from normalised logCPM values, used for computing single sample perturbation scores. Since genes with smaller logCPM turn to have a larger variance among single sample logFCs. A lowess curve is fitted to estimate the relationship between variance of single sample logFCs and mean of logCPM, and the relationship is used to estimate the variance of each mean logCPM value. Gene-wise weights, which are inverse of variances, are then multiplied to single sample logFCs to downweight genes with low counts.

It is assumed that the genes with extremely low counts have been removed and the count matrix has been normalised prior to logCPM matrix was derived. Rownames of the matrix must be genes' entrez ID. To convert other gene identifiers to entrz ID, see example.

Sample metadata should have the same number of rows as the number of columns in the logCPM matrix. Metadata also must have a column called "sample" storing sample names (column names of logCPM matrix), and a column called "treatment" storing treatment of each sample.The control treatment level specified by control parameter must exist in the treatment column.

This analysis was designed for experimental designs that include matched pairs of samples, such as when tissues collected from the same patient were treated with different treatments to study different treatment effects. Parameter factor tells the function how samples can be put into matching pairs. It must also be included as a column in the metadata.

Examples

require(AnnotationHub)
require(ensembldb)
# convert rownamews of logCPM from gene ids to gene entrez IDs through `AnnotationHub`
ah <- AnnotationHub()
#> snapshotDate(): 2021-10-20
ah <- subset(ah,genome == "GRCh38" & title == "Ensembl 101 EnsDb for Homo sapiens")
ensDb <- ah[[1]]
#> loading from cache
rownames(logCPM_example) <- mapIds(ensDb, rownames(logCPM_example), "ENTREZID", keytype = "GENEID")
#> Warning: Unable to map 3137 of 18481 requested IDs.

# Remove genes that couldn't be matched to entrez IDs
logCPM_example <- logCPM_example[!is.na(rownames(logCPM_example)),]

# Inspect metadata data frame to make sure it has treatment, sample and patient columns
head(metadata_example)
#> # A tibble: 6 × 4
#>   patient treatment sample         PR      
#>   <fct>   <fct>     <fct>          <fct>   
#> 1 N2      Vehicle   Vehicle_N2_48  Negative
#> 2 N2      E2+R5020  E2+R5020_N2_48 Negative
#> 3 N2      R5020     R5020_N2_48    Negative
#> 4 N3      Vehicle   Vehicle_N3_48  Negative
#> 5 N3      E2+R5020  E2+R5020_N3_48 Negative
#> 6 N3      R5020     R5020_N3_48    Negative
length(setdiff(colnames(logCPM_example), metadata_example$sample)) == 0
#> [1] TRUE
ls <- weight_ssFC(logCPM_example, metadata = metadata_example,
 factor = "patient", control = "Vehicle")